Add AGPL license Since I have very likely rewritten/architected the majority of this app to now fit a different purpose (multi club use) rather than the one it was originally built for (single club use), I feel reasonably safe applying this license.
If anyone has issues with specific parts of the code, let me know and I'll do my best to make sure everyone's rights and feelings are as respected as possible.

I have informed max about my intent to add this license in early September, 2025 via an email that was responded to.